requirements.txt
numpy==2.2.3
opencv-contrib-python-headless==4.11.0.86
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyYAML==6.0.2

